src/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic-settings==2.2.1
motor==3.6.1
langchain==0.1.20
pypdf==5.3.0
chromadb==0.5.23
pytest==8.3.4
openai==1.65.1
requests==2.32.3
pytest-mock==3.14.0
tiktoken==0.9.0
streamlit==1.42.2
plotly==6.0.0
Werkzeug==3.1.3
nltk==3.9.1
sentence-transformers==3.4.1
matplotlib==3.9.4
requests==2.32.3
pdf2image==1.17.0
pillow==11.1.0
fitz==0.0.1.dev2
PyMuPDF==1.25.4
streamlit-drawable-canvas==0.9.3
pdfplumber==0.11.5
opencv-python==4.11.0.86
streamlit-ace==0.1.1
pytesseract==0.3.13
streamlit-javascript==0.1.5